# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ETF Weekly Report API",
    "description": "API for Korean ETF analysis and reporting",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/data/collect-all": {
      "post": {
        "tags": [
          "Data Collection"
        ],
        "summary": "Collect All Data",
        "description": "전체 종목 데이터 일괄 수집\n\n등록된 모든 종목(6개)의 가격 데이터와 매매동향을 일괄 수집합니다.\n\n**Query Parameters:**\n- days: 수집할 일수 (기본: 1일, 최대: 365일)\n\n**Example Request:**\n```\n# 당일 데이터만 수집 (기본)\nPOST /api/data/collect-all\n\n# 최근 10일 데이터 수집\nPOST /api/data/collect-all?days=10\n```\n\n**Example Response:**\n```json\n{\n  \"message\": \"Data collection completed for 6 tickers\",\n  \"result\": {\n    \"total_tickers\": 6,\n    \"success_count\": 6,\n    \"fail_count\": 0,\n    \"total_price_records\": 6,\n    \"total_trading_flow_records\": 6,\n    \"total_news_records\": 0,\n    \"details\": {\n      \"487240\": {\n        \"name\": \"삼성 KODEX AI전력핵심설비 ETF\",\n        \"success\": true,\n        \"price_records\": 1,\n        \"trading_flow_records\": 1,\n        \"news_records\": 0\n      }\n    }\n  }\n}\n```\n\n**Status Codes:**\n- 200: 성공\n- 400: 잘못된 파라미터\n- 503: 데이터 소스 일시적 오류\n- 500: 서버 오류\n\n**Notes:**\n- 일반적으로 스케줄러가 자동으로 실행하므로 수동 호출은 불필요\n- 데이터 갱신이 필요한 경우에만 사용\n- 대량 수집 시 시간이 오래 걸릴 수 있음 (약 6초/종목)",
        "operationId": "collect_all_data_api_data_collect_all_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "description": "수집할 일수 (기본: 1일)",
              "default": 1,
              "title": "Days"
            },
            "description": "수집할 일수 (기본: 1일)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/backfill": {
      "post": {
        "tags": [
          "Data Collection"
        ],
        "summary": "Backfill Data",
        "description": "모든 종목의 히스토리 데이터를 백필\n\n- **days**: 백필할 일수 (1-{MAX_COLLECTION_DAYS}일, 기본: {DEFAULT_BACKFILL_DAYS}일)\n\nReturns:\n    백필 결과 및 종목별 상세 정보",
        "operationId": "backfill_data_api_data_backfill_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "description": "백필할 일수 (기본: 90일)",
              "default": 90,
              "title": "Days"
            },
            "description": "백필할 일수 (기본: 90일)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/data/reset": {
      "delete": {
        "tags": [
          "Data Collection"
        ],
        "summary": "Reset Database",
        "description": "데이터베이스 초기화 (위험!)\n\nprices, news, trading_flow, collection_status, intraday_prices 테이블의 모든 데이터를 삭제합니다.\netfs 테이블은 유지됩니다.\nSQLite의 경우 sqlite_sequence 테이블도 초기화하여 AUTOINCREMENT ID를 1부터 다시 시작하도록 합니다.\n\n**⚠️ 경고: 이 작업은 되돌릴 수 없습니다!**\n\n**Example Response:**\n```json\n{\n  \"message\": \"Database reset successfully\",\n  \"deleted\": {\n    \"prices\": 1500,\n    \"news\": 250,\n    \"trading_flow\": 180,\n    \"collection_status\": 6,\n    \"intraday_prices\": 5000\n  }\n}\n```\n\n**Status Codes:**\n- 200: 성공\n- 500: 서버 오류",
        "operationId": "reset_database_api_data_reset_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}